Refactoring to add Versionable namespace
