+++ test/Versionable/Ferret/Detector/FileinfoTest.php
@@ -1,21 +1,25 @@
 <?php
-namespace Ferret\Detector;
+namespace Versionable\Ferret\Detector;
 
-require_once 'src/Ferret/Detector/DetectorInterface.php';
-require_once 'src/Ferret/Detector/DetectorAbstract.php';
-require_once 'src/Ferret/Detector/Pathinfo.php';
+require_once 'src/Versionabe/Ferret/Detector/DetectorInterface.php';
+require_once 'src/Versionabe/Ferret/Detector/DetectorAbstract.php';
+require_once 'src/Versionabe/Ferret/Detector/Fileinfo.php';
+
+require_once 'src/Versionabe/Ferret/Detector/Exception/DetectorException.php';
 
 
 /**
- * Test class for Pathinfo.
+ * Test class for Fileinfo.
  * Generated by PHPUnit on 2010-10-09 at 20:06:58.
  * 
  * @group Ferret
+ * @group ExternalLibrary
+ * @group Fileinfo
  */
-class PathinfoTest extends \PHPUnit_Framework_TestCase
+class FileinfoTest extends \PHPUnit_Framework_TestCase
 {
     /**
-     * @var Pathinfo
+     * @var Fileinfo
      */
     protected $object;
 
@@ -25,8 +29,7 @@ class PathinfoTest extends \PHPUnit_Framework_TestCase
      */
     protected function setUp()
     {
-      
-      $this->object = new Pathinfo;
+      $this->object = new Fileinfo;
     }
 
     /**
@@ -37,25 +40,21 @@ class PathinfoTest extends \PHPUnit_Framework_TestCase
     {
     }
 
+    /**
+     * @todo Implement testDetect().
+     */
     public function testDetect()
     {
       $this->assertEquals('text/plain', $this->object->detect('data/unit.txt'));
       
     }
     
+    /**
+     * @todo Implement testDetect().
+     */
     public function testDetectFail()
     {
       $this->assertFalse($this->object->detect('data/unit.false'));
       
-    } 
-    
-    public function testSetMapping()
-    {
-      $this->assertTrue($this->object->setMapping(array('txt' => 'text/plain')));
-    }
-
-    public function testSetMappingFail()
-    {
-      $this->assertFalse($this->object->setMapping(null));
-    }   
+    }    
 }
