+++ test/Versionable/Ferret/Detector/MIMETypeTest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Ferret\Detector;
+namespace Versionable\Ferret\Detector;
 
-require_once 'src/Ferret/Detector/DetectorInterface.php';
-require_once 'src/Ferret/Detector/DetectorAbstract.php';
-require_once 'src/Ferret/Detector/MIMEType.php';
+require_once 'src/Versionabe/Ferret/Detector/DetectorInterface.php';
+require_once 'src/Versionabe/Ferret/Detector/DetectorAbstract.php';
+require_once 'src/Versionabe/Ferret/Detector/MIMEType.php';
 
 /**
  * Test class for MIMEType.
