+++ test/Versionable/Ferret/Detector/PathinfoTest.php
@@ -1,25 +1,21 @@
 <?php
-namespace Ferret\Detector;
+namespace Versionable\Ferret\Detector;
 
-require_once 'src/Ferret/Detector/DetectorInterface.php';
-require_once 'src/Ferret/Detector/DetectorAbstract.php';
-require_once 'src/Ferret/Detector/Fileinfo.php';
-
-require_once 'src/Ferret/Detector/Exception/DetectorException.php';
+require_once 'src/Versionabe/Ferret/Detector/DetectorInterface.php';
+require_once 'src/Versionabe/Ferret/Detector/DetectorAbstract.php';
+require_once 'src/Versionabe/Ferret/Detector/Pathinfo.php';
 
 
 /**
- * Test class for Fileinfo.
+ * Test class for Pathinfo.
  * Generated by PHPUnit on 2010-10-09 at 20:06:58.
  * 
  * @group Ferret
- * @group ExternalLibrary
- * @group Fileinfo
  */
-class FileinfoTest extends \PHPUnit_Framework_TestCase
+class PathinfoTest extends \PHPUnit_Framework_TestCase
 {
     /**
-     * @var Fileinfo
+     * @var Pathinfo
      */
     protected $object;
 
@@ -29,7 +25,8 @@ class FileinfoTest extends \PHPUnit_Framework_TestCase
      */
     protected function setUp()
     {
-      $this->object = new Fileinfo;
+      
+      $this->object = new Pathinfo;
     }
 
     /**
@@ -40,21 +37,25 @@ class FileinfoTest extends \PHPUnit_Framework_TestCase
     {
     }
 
-    /**
-     * @todo Implement testDetect().
-     */
     public function testDetect()
     {
       $this->assertEquals('text/plain', $this->object->detect('data/unit.txt'));
       
     }
     
-    /**
-     * @todo Implement testDetect().
-     */
     public function testDetectFail()
     {
       $this->assertFalse($this->object->detect('data/unit.false'));
       
-    }    
+    } 
+    
+    public function testSetMapping()
+    {
+      $this->assertTrue($this->object->setMapping(array('txt' => 'text/plain')));
+    }
+
+    public function testSetMappingFail()
+    {
+      $this->assertFalse($this->object->setMapping(null));
+    }   
 }
